Add new DB as early as possible + link to new DB
This (mostly kinda sorta) fixes a race condition triggered by
adding a Datasette Database in the middle of a view. Views
trigger refreshes of DBs, which can explode when we add a new
DB to the dictionary of databases at the end of a view. Putting
this ahead of other (slower) features gives us a chance to avoid
exposing this flaw, but doesn't totally prevent it.

The worst that happens when the race condition is triggered is
the DB *is* imported, but the import says is failed.